fix: expose progress events in dial/dialProtocol types
Accept the `onProgress` event via types.
