+++ packages/transport-webrtc/src/private-to-public/options.ts
@@ -1,4 +1,4 @@
-import type { CreateListenerOptions, DialOptions } from '@libp2p/interface'
+import type { CreateListenerOptions, DialTransportOptions } from '@libp2p/interface'
 
 export interface WebRTCListenerOptions extends CreateListenerOptions {}
-export interface WebRTCDialOptions extends DialOptions {}
+export interface WebRTCDialOptions extends DialTransportOptions {}
